pyproject.toml
[project]
name = "text2sql-w-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "google-genai>=1.18.0",
    "httpx>=0.28.1",
    "instructor[claude]>=1.8.3",
    "mcp[cli]>=1.9.1",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "structlog>=25.3.0",
]

[dependency-groups]
dev = [
    "coverage>=7.2.0",
    "ipykernel>=6.29.5",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]
